Logs/metagpt/part3-2/instruction15/quiz_automation_mac_1766905667/requirements.txt
pytesseract>=0.3.10
Pillow>=9.0.0
pynput>=1.7.6
numpy>=1.21.0
pyobjc-framework-Quartz>=9.2